+++ sdk/include/opentelemetry/sdk/trace/batch_span_processor.h
@@ -139,7 +139,7 @@ private:
 
   /* Synchronization primitives */
   std::condition_variable cv_, force_flush_cv_;
-  std::mutex cv_m_, force_flush_cv_m_;
+  std::mutex cv_m_, force_flush_cv_m_, shutdown_m_;
 
   /* The buffer/queue to which the ended spans are added */
   common::CircularBuffer<Recordable> buffer_;